Improved login and sighup tests, renamed "test-options" to "fixtures" and updated playwright.config.ts
