+++ page-objects/companyPage.ts
@@ -9,18 +9,9 @@ export class CompanyPage {
   }
 
   async verifyWelcomeMessage(firstName: string, lastName: string) {
-    await expect(this.page.getByText(`Welcome ${firstName} ${lastName}`)).toBeVisible();
+    await expect(this.page.locator("[data-testid='welcome-header']")).toHaveText(`Welcome ${firstName} ${lastName}`);
   }
-
-  async verifyUserInLocalStorage(email: string) {
-    const users = await this.page.evaluate(() => {
-      const raw = localStorage.getItem('users');
-      return raw ? JSON.parse(raw) : { users: [] };
-    });
-
-    expect(users.users.some((u: any) => u.email === email)).toBe(true);
-  }
-
+  
   async clickLogout() {
     await this.page.getByRole('button', { name: 'Log out' }).click()
   }
